### 2023-09-25 #5864 @warrickball
Hi @MichielsenM! I'll be handling this submission as editor and will start looking for reviewers. In the meantime, I'd suggest having a look at JOSS's [review criteria](https://joss.readthedocs.io/en/latest/review_criteria.html) and [review checklist](https://joss.readthedocs.io/en/latest/review_checklist.html) in anticipation of what the reviewers will be looking for.

As it is, my first comment is that the paper is probably *too* high level, and doesn't give me a good idea of what the code does. At the moment it seems to rely on MESA and GYRE, so you should cite them.

The statement of need is also very scientific, rather than reviewing what software exists, or what it is about comparing models to data that's better than, say, something someone might write for themselves in a few hundred lines of Python. It's fair enough to say that there's simply no software that does this in a way that overcomes some quirk of g-mode asteroseismology. E.g. for solar-like oscillations we have [AIMS](https://ui.adsabs.harvard.edu/abs/2019MNRAS.484..771R) or [BASTA](https://ui.adsabs.harvard.edu/abs/2022MNRAS.509.4344A). Why can't they be repurposed for g-mode asteroseismology? I'm not claiming that they *can* but it's the sort of question JOSS papers should answer through the statement of need.
