+++ paper/paper.bib
@@ -16,6 +16,22 @@
 	adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
+@ARTICLE{Michielsen2023,
+       author = {{Michielsen}, Mathias and {Van Reeth}, Timothy and {Tkachenko}, Andrew and {Aerts}, Conny},
+        title = "{Probing the physics in the core boundary layers of the double-lined B-type binary KIC4930889 from its gravito-inertial modes}",
+      journal = {arXiv e-prints},
+     keywords = {Astrophysics - Solar and Stellar Astrophysics},
+         year = 2023,
+        month = sep,
+          eid = {arXiv:2309.13123},
+        pages = {arXiv:2309.13123},
+archivePrefix = {arXiv},
+       eprint = {2309.13123},
+ primaryClass = {astro-ph.SR},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2023arXiv230913123M},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
 @ARTICLE{Waelkens1991,
 	author = {{Waelkens}, C.},
 	title = "{Slowly pulsating B stars.}",
