### 2023-09-27 #5864 @MichielsenM
Hi @warrickball, thanks for pointing this out. I tried to write it very high level, but maybe I overdid it a bit.
The current setup indeed relies on MESA, but that's only in the initial step where it reads in theoretical models. If this step is adjusted to read in other file formats, the rest should in principle work with models from codes other than MESA. I had put this somewhere in the documentation, but maybe you're right and I should mention/cite this in the paper itself as well. I'll adjust the text accordingly.
