+++ docs/Walkthrough.md
@@ -218,7 +218,7 @@ The folder `KIC7760680` contains an example of a file with observational data (<
 
 If you calculated your own theoretical grids, you should set their path properly in the pipeline files with the function argument `grid_parent_directory`. However, for this walkthrough the folder `grid_summary` already contains the outputs of step 0 of the modelling pipeline. You can unpack these compressed files using terminal commands `tar -xf surfaceGrid_DO.tar.xz` and `tar -xf pulsationGrid_DO_rot0.4805_k0m1.tar.xz`.
 The theoretical grid summary that is included here is the radiative grid from <a href="https://doi.org/10.1051/0004-6361/202039926" target="_blank"> Michielsen et al. (2021)</a>. If you were to run the pipelines, you would find similar results to the paper, but slight differences would be present. These differences occur since the near-core rotation rate of the star was assumed to be fixed in the paper. Meanwhile the modelling pipeline has undergone numerous updates, one of which is the optimisation of the stellar rotation rate for each theoretical equilibrium model separately. 
-This optimisation is explained in more detail in Michielsen et al. (2023) (currently under revision).
+This optimisation is explained in more detail in <a href="https://arxiv.org/abs/2309.13123" target="_blank"> Michielsen et al. (2023)</a>.
 
 Starting the modelling with this pre-made setup is now as easy as running `python pipeline.py` from your terminal (after making sure the right python environment is activated).
 Note that you will see warnings `WARNING  file already existed: ../grid_summary/surfaceGrid_DO.hdf` and `WARNING  file already existed: ../grid_summary/pulsationGrid_DO_rot0.4805_k0m1.hdf`. This is normal since the code detects that the summary files are already present, and don't need to be extracted a second time. Step 0 of the modelling pipeline will therefore skip creating these files and notify you that they already existed. 
