+++ paper/paper.bib
@@ -17,7 +17,7 @@
 }
 
 @ARTICLE{Michielsen2023,
-       author = {{Michielsen}, Mathias and {Van Reeth}, Timothy and {Tkachenko}, Andrew and {Aerts}, Conny},
+       author = {{Michielsen}, M. and {Van Reeth}, T. and {Tkachenko}, A. and {Aerts}, C.},
         title = "{Probing the physics in the core boundary layers of the double-lined B-type binary KIC4930889 from its gravito-inertial modes}",
       journal = {arXiv e-prints},
      keywords = {Astrophysics - Solar and Stellar Astrophysics},
@@ -46,8 +46,8 @@ archivePrefix = {arXiv},
 }
 
 @BOOK{Aerts2010,
-	author = {{Aerts}, Conny and {Christensen-Dalsgaard}, J{\o}rgen and
-	{Kurtz}, Donald W.},
+	author = {{Aerts}, C. and {Christensen-Dalsgaard}, J. and
+	{Kurtz}, D. W.},
 	title = "{Asteroseismology}",
 	booktitle = {Asteroseismology, Astronomy and Astrophysics Library. ISBN 978-1-4020-5178-4. Springer Science+Business Media B.V., 2010, p.},
 	year = "2010",
@@ -90,6 +90,251 @@ archivePrefix = {arXiv},
        eprint = {1912.12300},
  primaryClass = {astro-ph.SR},
        adsurl = {https://ui.adsabs.harvard.edu/abs/2021RvMP...93a5001A},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+@ARTICLE{AguirreBorsenKoch2022,
+       author = {{Aguirre B{\o}rsen-Koch}, V. and {R{\o}rsted}, J.~L. and {Justesen}, A.~B. and {Stokholm}, A. and {Verma}, K. and {Winther}, M.~L. and {Knudstrup}, E. and {Nielsen}, K.~B. and {Sahlholdt}, C. and {Larsen}, J.~R. and {Cassisi}, S. and {Serenelli}, A.~M. and {Casagrande}, L. and {Christensen-Dalsgaard}, J. and {Davies}, G.~R. and {Ferguson}, J.~W. and {Lund}, M.~N. and {Weiss}, A. and {White}, T.~R.},
+        title = "{The BAyesian STellar algorithm (BASTA): a fitting tool for stellar studies, asteroseismology, exoplanets, and Galactic archaeology}",
+      journal = {\mnras},
+     keywords = {asteroseismology, methods: numerical, methods: statistical, stars: fundamental parameters, Astrophysics - Solar and Stellar Astrophysics, Astrophysics - Earth and Planetary Astrophysics, Astrophysics - Astrophysics of Galaxies},
+         year = 2022,
+        month = jan,
+       volume = {509},
+       number = {3},
+        pages = {4344-4364},
+          doi = {10.1093/mnras/stab2911},
+archivePrefix = {arXiv},
+       eprint = {2109.14622},
+ primaryClass = {astro-ph.SR},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2022MNRAS.509.4344A},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+@ARTICLE{Rendle2019,
[…]
