### 2023-09-27 #5864 @MichielsenM
I've moved the scientific part to an introduction section, and added a new statement of need, which I hope is more in line with what that section should contain. Also, would an "overview" section where I describe the steps in the workflow be a good approach to make it less high level?
