+++ tests/test_build_optimised_pattern.py
@@ -1,20 +1,67 @@
+import numpy as np
 from foam import build_optimised_pattern as bop
 
 def test_generate_spacing_series():
-    """ test period spacing series calculated by generate_spacing_series"""
+    """Test period spacing series calculated by generate_spacing_series"""
     periods = [0.2, 0.3, 0.38, 0.42, 0.45]
     expected = ([0.1*86400, 0.08*86400, 0.04*86400, 0.03*86400], None)
     result = bop.generate_spacing_series(periods)
 
-    assert all ([(a-b)/a< 1E-10 for a,b in zip(expected[0], result[0])])
+    assert all ([abs((a-b)/a)< 1E-10 for a,b in zip(expected[0], result[0])])
     assert result[1] == expected[1]
 
 def test_generate_spacing_series_with_errors():
-    """ test period spacing series and its errors calulated by test_generate_spacing_series """
+    """Test period spacing series and its errors calulated by test_generate_spacing_series"""
     periods = [1.1, 1.3, 1.39, 1.47, 1.52 ]
-    errors = [1E-6, 2E-7, 6E-6, 4E-6, 2E-6]
-    expected = ([0.2*86400, 0.09*86400, 0.08*86400, 0.05*86400], [8.9856E-8*86400, 3.113856E-6*86400, 4.4928E-6*86400, 1.728E-8*86400])
+    errors = [1E-1, 2E-1, 4E-1, 3E-1, 2E-1]
+    expected_spacing = [0.2*86400, 0.09*86400, 0.08*86400, 0.05*86400]
+    expected_errors  = [np.sqrt(0.05)*86400,  np.sqrt(0.2)*86400, np.sqrt(0.25)*86400, np.sqrt(0.13)*86400]
     result = bop.generate_spacing_series(periods, errors)
 
-    assert all ([(a-b)/a< 1E-10 for a,b in zip(expected[0], result[0])])
-    assert all ([(a-b)/a< 1E-10 for a,b in zip(expected[1], result[1])])
+    assert all ([abs((a-b)/a)< 1E-10 for a,b in zip(expected_spacing, result[0])])
+    assert all ([abs((a-b)/a)< 1E-10 for a,b in zip(expected_errors, result[1])])
+
+def test_puls_series_from_given_puls_at_end():
+    """Test building theoretical pulsation pattern from end of pattern."""
+    Theo_freq = np.asarray([0.6, 0.7, 0.8, 0.9, 1.0, 1.1, 1.2, 1.3])
+    Obs_freq  = np.asarray([0.68, 0.76, 0.88, 0.95, 1.02, 1.06])
+    Obs_to_build_from = 1.06
+    expected = [0.6, 0.7, 0.8, 0.9, 1.0, 1.1]
+    result = bop.puls_series_from_given_puls(Theo_freq, Obs_freq, Obs_to_build_from)
+
+    assert all ([abs((a-b)/a)< 1E-10 for a,b in zip(expected, result)])
+
+def test_puls_series_from_given_puls_midway():
+    """Test building theoretical pulsation pattern from midway pattern."""
+    Theo_freq = np.asarray([0.6, 0.7, 0.8, 0.9, 1.0, 1.1, 1.2, 1.3])
+    Obs_freq = np.asarray([0.68, 0.76, 0.88, 0.95, 1.02, 1.06])
+    Obs_to_build_from = 0.88
+    expected = [0.7, 0.8, 0.9, 1.0, 1.1, 1.2]
+    result = bop.puls_series_from_given_puls(Theo_freq, Obs_freq, Obs_to_build_from)
[…]
