### 2023-09-27 #5864 @warrickball
> Also, would an "overview" section where I describe the steps in the workflow be a good approach to make it less high level?

Yes, that would be good. As it is, I don't think the paper describes what the code actually does, and a few lines on that would help.

Don't put too much weight on my remarks yet because the reviewers may have comments that overlap or lead to further changes. I've actually managed to find two reviewers surprisingly quickly. In a moment I'll start the actual review, where any conversations can continue.
