### 2023-09-27 #5884 @editorialbot
[REVIEW]: Foam: A Python package for forward asteroseismic modelling of gravity modes
**Submitting author:** @MichielsenM (<a href="http://orcid.org/0000-0001-9097-3655">Mathias Michielsen</a>)
**Repository:** https://github.com/MichielsenM/FOAM
**Branch with paper.md** (empty if default branch):
**Version:** V1.0.0
