### 2023-09-27 cda0026
Clarifying some options
