+++ docs/Walkthrough.md
@@ -221,7 +221,7 @@ The theoretical grid summary that is included here is the radiative grid from <a
 This optimisation is explained in more detail in <a href="https://arxiv.org/abs/2309.13123" target="_blank"> Michielsen et al. (2023)</a>.
 
 Starting the modelling with this pre-made setup is now as easy as running `python pipeline.py` from your terminal (after making sure the right python environment is activated).
-Note that you will see warnings `WARNING  file already existed: ../grid_summary/surfaceGrid_DO.hdf` and `WARNING  file already existed: ../grid_summary/pulsationGrid_DO_rot0.4805_k0m1.hdf`. This is normal since the code detects that the summary files are already present, and don't need to be extracted a second time. Step 0 of the modelling pipeline will therefore skip creating these files and notify you that they already existed. If you want to speed up running the example setup, you can add `pattern_methods = ['highest_amplitude', 'highest_frequency']` to the pipeline configuration options. This will then enable only 2 out of the 3 options for constructing the theoretical pulsation patterns. The method `chisq_longest_sequence` will be left out, since this one is far slower than the others and will add a considerable amount of time to step 1 of the pipeline. It will of course remove some of the time in the other steps as well. Alternatively, you can also increase `nr_cpu` depending on your computational setup.
+Note that you will see warnings `WARNING  file already existed: ../grid_summary/surfaceGrid_DO.hdf` and `WARNING  file already existed: ../grid_summary/pulsationGrid_DO_rot0.4805_k0m1.hdf`. This is normal since the code detects that the summary files are already present, and don't need to be extracted a second time. Step 0 of the modelling pipeline will therefore skip creating these files and notify you that they already existed. If you want to speed up running the example setup, you can add `pattern_methods = ['highest-amplitude', 'highest-frequency']` to the pipeline configuration options. This will then enable only 2 out of the 3 options for constructing the theoretical pulsation patterns. The method `chisq_longest_sequence` will be left out, since this one is far slower than the others and will add a considerable amount of time to step 1 of the pipeline. It will of course remove some of the time in the other steps as well. Alternatively, you can also increase `nr_cpu` depending on your computational setup.
 
 As explained before, the expected results are slightly different from those published by <a href="https://doi.org/10.1051/0004-6361/202039926" target="_blank"> Michielsen et al. (2021)</a>. The expected results of this pipeline are therefore included in directory <a href="https://github.com/MichielsenM/FOAM/tree/master/example_setup/expected_outcome" target="_blank"> expected_outcome</a>. This folder contains the results for modelling the full grid and the nested subgrid. It contains the output of the last step (step 7), which is a table with the best model parameters for each combination of the chosen theoretical grid, seismic observables, and pattern construction methods. One table holds these for a chi-squared (CS) merit function, whilst the other holds these for a Mahalanobis Distance (MD) merit function. Furthermore some of the figures created by step 6 of the pipeline have been included as well. (See [Pipeline modules](./Pipeline.md) for more information on the output of the different steps in the modelling.)
 
