### 2023-09-28 0af8cad
Use provided grid_parameters when extracting grids
