### 2023-09-28 c0f3b2f
hyperlinks to examples
