### 2023-10-14 3529f9d
Clarify docs + Rename highest-amp. to given-puls.
