+++ foam/build_optimised_pattern.py
@@ -37,7 +37,7 @@ def generate_spacing_series(periods, errors=None):
     return spacings, spacings_errors
 
 ################################################################################
-def construct_theoretical_puls_pattern(pulsationGrid_file, observations_file, method_build_series, highest_amplitude_pulsation=[], which_observable='period',
+def construct_theoretical_puls_pattern(pulsationGrid_file, observations_file, method_build_series, pattern_starting_pulsation=[], which_observable='period',
                                         output_file=f'theoretical_frequency_patterns.hdf', asymptotic_object=None, estimated_rotation=None, grid_parameters=None, nr_cpu=None):
     """
     Construct the theoretical frequency pattern for each model in the grid, which correspond to the observed pattern.
@@ -52,12 +52,12 @@ def construct_theoretical_puls_pattern(pulsationGrid_file, observations_file, me
         Column names specify the observable, and "_err" suffix denotes that it's the error.
     method_build_series: string
         way to generate the theoretical frequency pattern from each model to match the observed pattern. Options are:
-            highest-amplitude: build pattern from the observed highest amplitude    (function 'puls_series_from_given_puls')
+            provided-pulsation: build pattern from the provided pulsation    (function 'puls_series_from_given_puls')
             highest-frequency: build pattern from the observed highest frequency    (function 'puls_series_from_given_puls')
             chisq-longest-sequence: build pattern based on longest, best matching sequence of pulsations (function 'chisq_longest_sequence')
-    highest_amplitude_pulsation: array of floats
-        Only needed if you set method_build_series=highest-amplitude
-        Value of the pulsation with the highest amplitude, one for each separated part of the pattern.
+    pattern_starting_pulsation: array of floats
+        Only needed if you set method_build_series=provided-pulsation
+        Value of the pulsation to start building the pattern from, one for each separated part of the pattern.
         The unit of this value needs to be the same as the observable set through which_observable.
     which_observable: string
         Observable used in the theoretical pattern construction, options are 'frequency' or 'period'.
@@ -81,7 +81,7 @@ def construct_theoretical_puls_pattern(pulsationGrid_file, observations_file, me
 
     # partial function fixes all parameters of the function except for 1 that is iterated over in the multiprocessing pool.
     theo_pattern_func = partial(theoretical_pattern_from_dfrow, Obs=Obs, ObsErr=ObsErr, which_observable=which_observable,
-                                method_build_series=method_build_series, highest_amp_puls=highest_amplitude_pulsation, grid_parameters=grid_parameters,
+                                method_build_series=method_build_series, pattern_starting_pulsation=pattern_starting_pulsation, grid_parameters=grid_parameters,
                                 asymptotic_object=asymptotic_object, estimated_rotation=estimated_rotation, plot_rotation_optimisation=False)
 
     Path(Path(output_file).parent).mkdir(parents=True, exist_ok=True)   # Make the output file directory
@@ -106,7 +106,7 @@ def construct_theoretical_puls_pattern(pulsationGrid_file, observations_file, me
 
 # %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
 def theoretical_pattern_from_dfrow(summary_grid_row, Obs, ObsErr, which_observable, method_build_series,
-    highest_amp_puls=[], asymptotic_object=None, estimated_rotation=None, plot_rotation_optimisation=False, grid_parameters=None):
[…]
