+++ foam/pipeline/pipe1_construct_pattern.py
@@ -23,7 +23,7 @@ for grid in config.grids:
             puls_file = f'extracted_freqs/{observed_quantity}_{config.star}_{grid}_{method}.hdf'
             if not Path(puls_file).is_file():
                 bop.construct_theoretical_puls_pattern(f'../grid_summary/pulsationGrid_{grid}_rot{config.rotation_gyre}_k{config.kval}m{config.mval}.hdf', config.observations,
-                                                    method, highest_amplitude_pulsation=config.highest_amplitude_pulsation[observed_quantity], which_observable=observed_quantity, output_file=puls_file,
+                                                    method, pattern_starting_pulsation=config.pattern_starting_pulsation[observed_quantity], which_observable=observed_quantity, output_file=puls_file,
                                                     asymptotic_object=asymp_obj, estimated_rotation=config.rotation_gyre, grid_parameters=config.grid_parameters, nr_cpu=config.nr_cpu)
             else:
                 config.logger.warning(f'file already existed: {puls_file}')

[…]
