### 2023-10-18 594f95c
updated example setup
