+++ example_setup/expected_outcome/NESTED-SUBGRID_KIC7760680_best-model-table_MD.txt
@@ -1,4 +1,4 @@
 Grid Observables Pattern_construction Z M logD Xc aov fov Omega_rot MD AICc_MD
 DO dP chisq-longest-sequence 0.019 3.3 1.0 0.42 0.0 0.0 0.464 73.85 -386.23
-DO dP highest-amplitude 0.019 3.2 0.5 0.42 0.0 0.0 0.466 73.0 -388.14
+DO dP provided-pulsation 0.019 3.2 0.5 0.42 0.0 0.0 0.466 73.0 -388.14
 DO dP highest-frequency 0.023 3.3 0.5 0.4 0.0 0.0 0.4711 68.59 -392.99
