### 2023-10-18 b38bf81
Bugfix regarding missing and isolated frequencies
