### 2023-10-23 a4b2e09
Pipeline logging messages
