### 2023-10-23 c72e8e6
isolated frequencies with period spacings
