+++ docs/Walkthrough.md
@@ -37,8 +37,7 @@ title: Walkthrough
  Furthermore, any additional observables that you want to include in the merit function can be added in this file.
 
  Frequencies are preferred to be listed decreasing in value, so increasing in period. The pipeline can handle the other way around if the theoretical grid is also extracted in that way. (If the frequencies are ordered decreasing in value, the highest frequency parts in an interrupted pattern will be matched first, which is preferred since those have a lower relative uncertainty). One or more missing frequencies in the observed pattern can be indicated by 'f_missing' as index and 0 values for the frequency, period, and their errors. This will set the pipeline to model it as an interrupted period spacing pattern. (One entry of 'f_missing' denotes an interruption in the pattern due to one or more missing frequencies.) 
- 
- Splitting the pattern in multiple parts that each hold more than one frequency works for the different setups. However, the occurence an isolated frequency in between missing frequencies (f_(n-1), f_missing, f_n, f_missing, f_(n+1), f_(n+2)...) should be avoided when period spacings are used as seismic observables (`observable_seismic=['dP']`, see [Pipeline configuration](./Configuration.md)), since they won't have a period spacing associated with them, and the numerical values for the statistical criteria will have an offset compared to their actual value (constant and equal for all models in the setup). Isolated frequencies pose no problem and can be used in case periods and/or frequencies are used as seismic observables (`observable_seismic=['P', 'f']`).
+Keep in mind the occurence of an isolated frequency in between missing frequencies (f_(n-1), f_missing, f_n, f_missing, f_(n+1), f_(n+2)...) will not be taken into account in the merit function when period spacings are used as seismic observables (`observable_seismic=['dP']`, see [Pipeline configuration](./Configuration.md)), since they won't have a period spacing associated with them.
 
 An example of what the structure of such a file with the observational data looks like is given in the table below.
 
