### 2023-10-23 4e34310
small performance improvement
