+++ foam/build_optimised_pattern.py
@@ -79,9 +79,17 @@ def construct_theoretical_puls_pattern(pulsationGrid_file, observations_file, me
     Obs    = np.asarray(Obs_dFrame[which_observable])
     ObsErr = np.asarray(Obs_dFrame[f'{which_observable}_err'])
 
+    missing_puls = np.where(Obs==0)[0]          # if frequency was filled in as 0, it indicates an interruption in the pattern
+    Obs_without_missing=Obs[Obs!=0]                             # remove values indicating interruptions in the pattern
+    ObsErr_without_missing=ObsErr[ObsErr!=0]                    # remove values indicating interruptions in the pattern
+    missing_puls=[ missing_puls[i]-i for i in range(len(missing_puls)) ]    # Ajust indices for removed 0-values of missing frequencies
+
+    Obs_pattern_parts = np.split(Obs_without_missing, missing_puls)    # split into different parts of the interrupted pattern
+    ObsErr_pattern_parts = np.split(ObsErr_without_missing, missing_puls)
+
     # partial function fixes all parameters of the function except for 1 that is iterated over in the multiprocessing pool.
-    theo_pattern_func = partial(theoretical_pattern_from_dfrow, Obs=Obs, ObsErr=ObsErr, which_observable=which_observable,
-                                method_build_series=method_build_series, pattern_starting_pulsation=pattern_starting_pulsation, grid_parameters=grid_parameters,
+    theo_pattern_func = partial(theoretical_pattern_from_dfrow, Obs=Obs, Obs_pattern_parts=Obs_pattern_parts, ObsErr_pattern_parts=ObsErr_pattern_parts,
+                                which_observable=which_observable, method_build_series=method_build_series, pattern_starting_pulsation=pattern_starting_pulsation, grid_parameters=grid_parameters,
                                 asymptotic_object=asymptotic_object, estimated_rotation=estimated_rotation, plot_rotation_optimisation=False)
 
     Path(Path(output_file).parent).mkdir(parents=True, exist_ok=True)   # Make the output file directory
@@ -105,7 +113,7 @@ def construct_theoretical_puls_pattern(pulsationGrid_file, observations_file, me
     df.to_hdf(f'{output_file}', 'selected_puls_grid', format='table', mode='w')
 
 # %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
-def theoretical_pattern_from_dfrow(summary_grid_row, Obs, ObsErr, which_observable, method_build_series,
+def theoretical_pattern_from_dfrow(summary_grid_row, Obs, Obs_pattern_parts, ObsErr_pattern_parts, which_observable, method_build_series,
     pattern_starting_pulsation=[], asymptotic_object=None, estimated_rotation=None, plot_rotation_optimisation=False, grid_parameters=None):
     """
     Extract model parameters and a theoretical pulsation pattern from a row of the dataFrame that contains all model parameters and pulsation frequencies.
@@ -115,8 +123,10 @@ def theoretical_pattern_from_dfrow(summary_grid_row, Obs, ObsErr, which_observab
         second tuple entry is a pandas series, containing a row from the pandas dataFrame. (This row holds model parameters and pulsation frequencies.)
     Obs: numpy array
         Array of observed frequencies or periods. (Ordered increasing in frequency.)
-    ObsErr: numpy array
-        Array of errors on the observed frequencies or periods.
+    Obs_pattern_parts, ObsErr_pattern_parts: list of numpy arrays
+        Holds a numpy array (with the observed frequencies or periods, or with their errors) per split off part
[…]
