### 2023-10-27 08aa10d
correct number observables for dP
