### 2023-10-27 f9e77f9
example setup missing frequencies
