+++ docs/Walkthrough.md
@@ -228,3 +228,6 @@ Note that you will see warnings `WARNING  file already existed: ../grid_summary/
 As explained before, the expected results are slightly different from those published by <a href="https://doi.org/10.1051/0004-6361/202039926" target="_blank"> Michielsen et al. (2021)</a>. The expected results of this pipeline are therefore included in directory <a href="https://github.com/MichielsenM/FOAM/tree/master/example_setup/expected_outcome" target="_blank"> expected_outcome</a>. This folder contains the results for modelling the full grid and the nested subgrid. It contains the output of the last step (step 7), which is a table with the best model parameters for each combination of the chosen theoretical grid, seismic observables, and pattern construction methods. One table holds these for a chi-squared (CS) merit function, whilst the other holds these for a Mahalanobis Distance (MD) merit function. Furthermore some of the figures created by step 6 of the pipeline have been included as well. (See [Pipeline modules](./Pipeline.md) for more information on the output of the different steps in the modelling.)
 
 
+### Example setup interrupted pattern
+Another example is provided for the case of an interrupted pattern with some frequencies missing.
+For this setup, we make a new list of frequencies, <a href="https://github.com/MichielsenM/FOAM/blob/master/example_setup/KIC7760680_missing_freq/data_KIC7760680_missing_freq.tsv" target="_blank"> data_KIC7760680_missing_freq.tsv </a>, where we presumed that frequencies f4, f6, f18, and f19 from the original list <a href="https://github.com/MichielsenM/FOAM/blob/master/example_setup/KIC7760680/data_KIC7760680.tsv" target="_blank"> data_KIC7760680.tsv </a> are now missing. Additionally, we removed the luminosity information from the data file. Some notable differences in the configuration of the pipeline are in the settings for `pattern_starting_pulsation`, `N_pattern_parts`, and the different number for `N_periods`. Although not necessary to run the pipeline, we changed `observable_seismic` from period spacings ('dP') to periods ('P') for demonstrative purposes. The expected results of this pipeline are included in directory <a href="https://github.com/MichielsenM/FOAM/tree/master/example_setup/expected_outcome_missing_freq" target="_blank"> expected_outcome_missing_freq</a>. The cornerplots that originally had an HR-diagram in the top right corner now show a Kiel diagram instead, since we removed the luminosity information from the observational data.
