### 2023-11-06 a8728e3
A&A DOI instead of Arxiv link
