+++ paper/paper.bib
@@ -18,19 +18,23 @@
 
 @ARTICLE{Michielsen2023,
        author = {{Michielsen}, M. and {Van Reeth}, T. and {Tkachenko}, A. and {Aerts}, C.},
-        title = "{Probing the physics in the core boundary layers of the double-lined B-type binary KIC4930889 from its gravito-inertial modes}",
-      journal = {arXiv e-prints},
-     keywords = {Astrophysics - Solar and Stellar Astrophysics},
+        title = "{Probing the physics in the core boundary layers of the double-lined B-type binary KIC 4930889 from its gravito-inertial modes}",
+      journal = {\aap},
+     keywords = {asteroseismology, stars: oscillations, convection, stars: interiors, methods: statistical, techniques: photometric, Astrophysics - Solar and Stellar Astrophysics},
          year = 2023,
-        month = sep,
-          eid = {arXiv:2309.13123},
-        pages = {arXiv:2309.13123},
+        month = nov,
+       volume = {679},
+          eid = {A6},
+        pages = {A6},
+          doi = {10.1051/0004-6361/202244192},
 archivePrefix = {arXiv},
        eprint = {2309.13123},
  primaryClass = {astro-ph.SR},
-       adsurl = {https://ui.adsabs.harvard.edu/abs/2023arXiv230913123M},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2023A&A...679A...6M},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
+
+
 
 @ARTICLE{Waelkens1991,
 	author = {{Waelkens}, C.},
