+++ docs/Index.md
@@ -7,7 +7,7 @@ permalink: /
 The main functionality of FOAM is to perform forwards asteroseismic modelling of gravity modes via a modelling pipeline.
 An example of a science case for which this pipeline was used is the modelling of the slowly pulsating B-type star KIC7760680 by <a href="https://doi.org/10.1051/0004-6361/202039926" target="_blank"> Michielsen et al. (2021)</a>.
 The paper used an early version of the pipeline without optimising the rotation for each model individually, but a lot of its functionality was already used in this paper.
-Numerous updates in both performance and functionality have been included since and are employed in the modelling of the B-type binary KIC4930889 by <a href="https://arxiv.org/abs/2309.13123" target="_blank"> Michielsen et al. (2023)</a>.
+Numerous updates in both performance and functionality have been included since and are employed in the modelling of the B-type binary KIC4930889 by <a href="https://doi.org/10.1051/0004-6361/202244192 " target="_blank"> Michielsen et al. (2023)</a>.
 
 There is also some additional functionality outside of the modelling pipeline, mostly in the form of plotting tools, check <a href="https://github.com/MichielsenM/FOAM/blob/master/foam/plot_tools.py" target="_blank"> plot_tools</a> for all available functions. An example to quickly make a Kippenhahn plot or Hertzsprung–Russell diagram from a given <a href="https://docs.mesastar.org/en/latest/index.html" target="_blank"> MESA</a> file in the following way:
 
