### 2023-11-28 #5884 @ashleychontos
So I'll happily get this rolling!

First off -- thank you @warrickball for inviting me to review the `FOAM` software package -- this is another exciting step forward for the asteroseismology community, which so far has mostly been a closed-door science. I've had a chance to read the manuscript and have some comments about that while I go through the software installation/documentation/usage.

### Manuscript
 - Overall I think the content, as is, is not quite there yet. I feel like it's missing an important piece of information, e.g., an example or more specific details about the package itself.
 - I think that a clear description of the functionality and purpose of the software has been provided but per JOSS recommendations of the **summary**, is not necessarily suited for a diverse, non-specialist audience.
 -  Some questions/comments to keep in mind during revisions:
    - The paper does not mention how the software is meant to be used (i.e. as purely a command-line tool or in a Jupyter notebook). I'm sure this will become more clear as I dig through the software but it's never a bad idea to mention this in the manuscript to keep it self-consistent.
    - Who is the intended userbase?
    - They sufficiently (and appropriately) state how `FOAM` compares to software that analyzes solar-like oscillations but given that it is a package to analyze and characterize gravity modes, I am missing its context with other similar packages. Is this truly the first-ever package to do this and if so, how can we be confident with the results? If not, how do results compare for different packages? All this is to say that the package could be better contextualized with the current **state of the field**.
    -  It is fine to point to other papers but things that are mentioned multiple times in the paper ought to be described in more detail for clarity. For example, merit functions were mentioned but are not well-motivated in the paper. For example, it unclear why a traditional chi-squared is not sufficient. Are results significantly impacted through the use of different merit functions? If so, does one systematically under/over estimate observed properties? etc. These are all important points that a user should know when using the `FOAM` package.
    - Avoid using jargon that is not sufficiently defined i.e. Mahalanobis distance.
