### 2023-11-28 #5884 @ashleychontos
@warrickball it looks like editorialbot is failing to run gitinspector for information on repo contributions. Do you know anything about this?
