### 2023-11-28 #5884 @ashleychontos
Some additional comments...

### Documentation

- The documentation could still use some work but there's already a lot of great content there. It is still unclear who the target audience is for the package. Does it work and is it tested across all platforms? (I couldn't tell in the tests)
- There are no community guidelines as required by JOSS.
- I noticed that the source code has a healthy amount of docstrings available, all which are nice and thorough. Therefore I was a bit surprised to notice that those didn't translate to the documentation, seems like a missed opportunity there for users wishing to become more advanced with `FOAM`.
- The first page is currently confusing! The mention of additional functionality straight away kinda blurs the main purpose and application of the software -- I suggest putting this in a more secondary location. Further, having links to source code (via plot_tools) should be avoided. This should be sufficiently documented without the user having to dig through several hundred lines of code to figure out additional plotting tools. At the very least, there can be a bulleted list in the documentation.
- While the required python packages are appropriately included in the repository, the dependencies are not clear stated in the documentation per JOSS standards.
- The documentation does not mention if there is a way to check if the installation is successful.
