### 2023-11-28 #5884 @warrickball
> @warrickball it looks like editorialbot is failing to run gitinspector for information on repo contributions. Do you know anything about this?

I've just had a look and this looks like an issue pretty deep in the JOSS toolchain. The output is meant to summarise who had contributed how much code. Here's what I got from running `gitinspector` locally (after applying a patch to make it work in Python ~3.8+):
```$ gitinspector -f** foam
Statistical information for the repository 'foam' was gathered on 2023/11/28.

JVB11                            1           113             46            0.82
Mathias Michielsen               6            38             11            0.25
MichielsenM                    247         12810           6331           98.56
Sarah Gebruers                   1            60             12            0.37

Below are the number of rows from each author that have survived and are still intact in the current revision:

Author                     Rows      Stability          Age       % in comments
Mathias Michielsen            1            2.6          6.4                0.00
MichielsenM                5963           46.5         11.2                6.09
Sarah Gebruers                4            6.7         10.8                0.00
```
I'm not sure what the "stability" or "age" metrics mean and couldn't easily find it in `gitinspector`'s docs.
