### 2023-11-29 #5884 @ashleychontos
This is perfect, thanks. I just wanted to make sure that the author list was properly representative of the contributed work.

> I'm not sure what the "stability" or "age" metrics mean and couldn't easily find it in `gitinspector`'s docs.

I think "age" actually corresponds to the length of time a given contribution (i.e. X rows) lasts in the repo, although the old age here would be confusing (if true). I double checked `pySYD`  and this seems to be the case i.e. the first commit was February 4, 2020 and my original JOSS submission was May 8, 2021, thus ~=1.2 years. I am less sure about what stability means in this context. In any case, the information you captured above is more than enough for what I needed it for. Thanks!
