### 2023-12-13 #5884 @warrickball
Hi @MichielsenM, I just thought I'd check in to see if you've noted the comments from @ashleychontos. You don't need to wait for further remarks (or @AnkitBarik); you can start working on Ashley's comments immediately. In this sense, a JOSS review is not like a normal article review. We like the process to be continuous and iterative.

I'm also aware that your default GitHub settings might mean you haven't been notified about anything other than the first message. (I've myself missed stuff on GitHub before because of unexpected defaults.) Let me know if you have any questions.
