### 2023-12-20 #5884 @warrickball
Thanks for letting me know. I don't think there's any reason not to proceed, even if more slowly. I'll still try to keep the (lower) momentum up, so you might find me checking in every few weeks to keep us inching forward.
