### 2023-12-20 #5884 @AnkitBarik
Hi @warrickball thank you for asking me to be a reviewer for this as well. Firstly, I apologize for getting swamped with a couple of conferences and deadlines and getting to this just now. I went through the conversation above and I agree with all the points that @ashleychontos made, so I won't fully repeat them again. ( Thank you Ashley for taking the lead on this. ) Here are some more comments from me:

### Manuscript

The manuscript was a bit difficult to understand for someone like me who deals with waves/modes in astrophysical bodies but is not deeply entrenched in asteroseismology and MESA/GYRE setups. This goes back to the point of the target audience. Even for a relatively expert audience, I would recommend detailing the manuscript with a brief section on methods and with a minimal example.

### Documentation

It is unclear from the documentation to what extent and where MESA/GYRE are needed for the package to function.
For example, the [prerequisites](https://michielsenm.github.io/FOAM/Installation) says ,

> Some functionality requires [GYRE](https://gyre.readthedocs.io/en/stable/) to be installed as well

but when I simply try to run the example pipeline, it doesn't run without GYRE being installed. If possible, can one have a minimal working example that doesn't require any additional installations? In addition, the documentation should provide what the expected output of that example would be, so the user can verify that the installation is working correctly.

Lastly, despite the above, the functions are fairly well documented and I do think the package merits publishing on JOSS after the above comments have been taken into account. I'll add more comments if something more comes to mind.

Sincerely,
Ankit
