+++ foam/build_optimised_pattern.py
@@ -42,7 +42,7 @@ def construct_theoretical_puls_pattern(pulsationGrid_file, observations_file, me
     """
     Construct the theoretical frequency pattern for each model in the grid, which correspond to the observed pattern.
     (Each theoretical model is a row in 'pulsationGrid_file'.)
-    The rotation rate will be scaled and optimased for each theoretical pattern individually. This optimisation will not be performed if asymptotic_object=None.
+    The rotation rate will be scaled and optimised for each theoretical pattern individually. This optimisation will not be performed if asymptotic_object=None.
     ------- Parameters -------
     pulsationGrid_file: string
         path to file containing input parameters of the models, and the pulsation frequencies of those models
@@ -82,7 +82,7 @@ def construct_theoretical_puls_pattern(pulsationGrid_file, observations_file, me
     missing_puls = np.where(Obs==0)[0]          # if frequency was filled in as 0, it indicates an interruption in the pattern
     Obs_without_missing=Obs[Obs!=0]                             # remove values indicating interruptions in the pattern
     ObsErr_without_missing=ObsErr[ObsErr!=0]                    # remove values indicating interruptions in the pattern
-    missing_puls=[ missing_puls[i]-i for i in range(len(missing_puls)) ]    # Ajust indices for removed 0-values of missing frequencies
+    missing_puls=[ missing_puls[i]-i for i in range(len(missing_puls)) ]    # Adjust indices for removed 0-values of missing frequencies
 
     Obs_pattern_parts = np.split(Obs_without_missing, missing_puls)    # split into different parts of the interrupted pattern
     ObsErr_pattern_parts = np.split(ObsErr_without_missing, missing_puls)
@@ -119,13 +119,13 @@ def theoretical_pattern_from_dfrow(summary_grid_row, Obs, Obs_pattern_parts, Obs
     Extract model parameters and a theoretical pulsation pattern from a row of the dataFrame that contains all model parameters and pulsation frequencies.
     ------- Parameters -------
     summary_grid_row: tuple, made of (int, pandas series)
-        tuple retruned from pandas.iterrows(), first tuple entry is the row index of the pandas dataFrame
+        tuple returned from pandas.iterrows(), first tuple entry is the row index of the pandas dataFrame
         second tuple entry is a pandas series, containing a row from the pandas dataFrame. (This row holds model parameters and pulsation frequencies.)
     Obs: numpy array
         Array of observed frequencies or periods. (Ordered increasing in frequency.)
     Obs_pattern_parts, ObsErr_pattern_parts: list of numpy arrays
         Holds a numpy array (with the observed frequencies or periods, or with their errors) per split off part
-        of the observerd pattern if it is interrupted.
+        of the observed pattern if it is interrupted.
         The list contains only one array if the observed pattern is uninterrupted.
     which_observable: string
         Which observables are used in the pattern building, options are 'frequency' or 'period'.
@@ -212,14 +212,14 @@ def theoretical_pattern_from_dfrow(summary_grid_row, Obs, Obs_pattern_parts, Obs
                 optimised_periods = optimised_pulsations
             
             # Recombine observational errors in one array
-            ComninedObsErr = []
+            CombinedObsErr = []
             for O_part in ObsErr_pattern_parts:
-                if (len(ComninedObsErr) > 0): 
-                    ComninedObsErr.extend([0])
-                ComninedObsErr.extend(O_part)
-            ComninedObsErr = np.asarray(ComninedObsErr)
[…]
