### 2024-01-04 9c668c3
Module interface to check if installation successful
