### 2024-01-04 766edb4
mention dependencies in installation instructions
