### 2024-02-18 f88ec05
Rewrite paper, add community guidelines, clarify docs
