+++ paper/paper.bib
@@ -342,3 +342,37 @@ archivePrefix = {arXiv},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
+@article{Eckart1960,
+	author = "Eckart, G.",
+	doi = "10.1002/qj.49708938224",
+	issn = "1477-870X",
+	journal = "Hydrodynamics of oceans and atmospheres, Pergamon Press, Oxford",
+	title = "{Hydrodynamics of oceans and atmospheres}",
+	url = "http://dx.doi.org/10.1002/qj.49708938224",
+	year = "1960"
+}
+
+
+@ARTICLE{Townsend2020,
+       author = {{Townsend}, R.~H.~D.},
+        title = "{Improved asymptotic expressions for the eigenvalues of Laplace's tidal equations}",
+      journal = {\mnras},
+     keywords = {hydrodynamics, waves, methods: analytical, methods: numerical, stars: oscillations, stars: rotation, Astrophysics - Solar and Stellar Astrophysics, Mathematical Physics},
+         year = 2020,
+        month = sep,
+       volume = {497},
+       number = {3},
+        pages = {2670-2679},
+          doi = {10.1093/mnras/staa2159},
+archivePrefix = {arXiv},
+       eprint = {2006.12596},
+ primaryClass = {astro-ph.SR},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2020MNRAS.497.2670T},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+@BOOK{Claeskens2008,
+	author = {{Claeskens}, G. and {Hjort}, N. L.},
+	title = "{Model Selection and Model Averaging, Cambridge Series in Statistical and Probabilistic Mathematics}",
+	year = 2008,
+}
