+++ docs/Guidelines.md
@@ -1,0 +1,28 @@
+---
+layout: default
+title: Community guidelines
+---
+# Vision
+
+Several tools have been developed for asteroseismic analyses, but many of them are closed-source and therefore inaccessible to the general astronomy community. Whereas some efforts have been made to provide publicly available open-source tools for asteroseismic modelling, these predominantly concern the analysis of solar-like oscillators. Foam tries to extend this effort to the modelling of gravity mode oscillators.
+
+# Contributing
+
+## Contribute to the software
+If you have features or other ideas you want to contribute to the code, please submit them as a pull request.
+If you're unsure about how to approach the implementation of your idea, feel free to contact [Mathias](mailto:michielsenmathias@gmail.com).
+
+## Report issues or bugs
+If you run into issues or bugs, you can [create an issue on github](https://github.com/MichielsenM/FOAM/issues). If you managed to find a solution, you're also welcome to create a pull request alongside the issue with a fix.
+
+## Questions or problems
+If you have specific questions or run into problems, you can contact [Mathias](mailto:michielsenmathias@gmail.com).
+
+
+## Collaborators
+
+Many thanks to these amazing collaborators that have helped with the development and/or testing of the software.
+
+Foam collaborators:
+- Timothy Van Reeth
+- Alex Kemp
