### 2024-02-18 #5884 @MichielsenM
@warrickball, @ashleychontos , @AnkitBarik, thank you for the feedback and your patience with me slowly working on this. I’ve pushed changes to the github repository to address your comments.

I rewrote the manuscript, and added a section ‘software package overview’ to the paper. I hope this will make the methods, functionality, and how it’s meant to be used clearer and (at least the summary) more suitable for a non-specialist audience.
As for the comment about similar packages, I (and my former collaborators) are unaware of other publicly available, open-source modelling packages for gravity mode asteroseismology. I can therefore only make the comparison to packages for modelling solar-like oscillations, and stress that this is an effort to extend the open-source modelling tools to the g-mode regime.

In the online documentation, I now included community guidelines, updated the installation to both have a check for success and more clarity on the dependencies, and made some small changes to clarify things.

@ashleychontos, you mentioned that it’s a missed opportunity that the healthy amount of docstrings didn't translate to the documentation. Could you be more specific about what is missing? I’m biased myself since I wrote it, and have a bit of a hard time to see what you’d like to be added.
