### 2024-02-18 cb3ea67
bulletpoints formatting
