+++ paper/paper.md
@@ -45,14 +45,15 @@ Some tools have been developed and made publicly available to model and determin
 The script to run the pipeline can be altered in order to change the modelling approach you want to take. The various configuration options, the installation procedure, and a walkthrough of how to create your own modelling setup, are described in more detail in the [online documentation](https://michielsenm.github.io/FOAM). Although it relies on grids of stellar equilibrium models computed by `MESA` as the source of the theoretical model grid, `MESA`'s installation itself is not required for `Foam` to function. The installation of `GYRE` is however required, specifically since `Foam` relies on the `tar_fit.mX.kX.h5` files included in the `GYRE` installation. This allows us to rescale the g-modes for various stellar rotation rates, following the traditional approximation of rotation [e.g. @Eckart1960, see @Townsend2020 for its implementation in `GYRE`] and assuming rigid rotation. This facilitates computing the oscillation frequencies for the grid of stellar equilibrium models only once, and subsequently rescaling them to find the optimised rotation rate [see @Michielsen2023]. This approach avoids repeating the oscillation computations for a variety of rotation values, which would introduce extra dimensionality in the modelling problem in the form of adding the rotation rate as an additional free parameter.
 
 `Foam`'s modelling procedure can be broken down into following sequential [steps of the pipeline](https://michielsenm.github.io/FOAM/Pipeline):
-- Extract all required parameters and quantities from the files in the theoretical `MESA` and `GYRE` grids.
-- Construct the theoretical pulsation patterns for each stellar model. Thereafter select theoretical pulsation patterns matching the observational pattern whilst optimising their rotation rates. Finally combine this information with the models' surface properties.
-- Calculate the likelihood of all the theoretical patterns according to the specified merit functions and observables. This list of observables consist of the pulsations, but can optionally be extended with spectroscopic or astrometric information.
-- Exclude all the models that fall outside an n-sigma error box on the spectroscopic and astrometric constraints as acceptable solutions.
-- Calculate the Akaike information criterion (AIC) [@Claeskens2008] corrected for small sample size. This statistical criterion rewards goodness of fit, but penalises model complexity in the form of additional free parameters. The AIC thus allows a statistical comparison between models of different (nested) grids where the number of free parameters is not the same.
-- Calculate the 2 sigma uncertainty region of the maximum likelihood solution using Bayes’ theorem.
-- Make corner plots for all combinations of the different modelling choices (See \autoref{fig:cornerplot} for an example).
-- Construct a table with the best model of the grid for each combination of different modelling choices.
+
+ - Extract all required parameters and quantities from the files in the theoretical `MESA` and `GYRE` grids.
+ - Construct the theoretical pulsation patterns for each stellar model. Thereafter select theoretical pulsation patterns matching the observational pattern whilst optimising their rotation rates. Finally combine this information with the models' surface properties.
+ - Calculate the likelihood of all the theoretical patterns according to the specified merit functions and observables. This list of observables consist of the pulsations, but can optionally be extended with spectroscopic or astrometric information.
+ - Exclude all the models that fall outside an n-sigma error box on the spectroscopic and astrometric constraints as acceptable solutions.
+ - Calculate the Akaike information criterion (AIC) [@Claeskens2008] corrected for small sample size. This statistical criterion rewards goodness of fit, but penalises model complexity in the form of additional free parameters. The AIC thus allows a statistical comparison between models of different (nested) grids where the number of free parameters is not the same.
+ - Calculate the 2 sigma uncertainty region of the maximum likelihood solution using Bayes’ theorem.
+ - Make corner plots for all combinations of the different modelling choices (See \autoref{fig:cornerplot} for an example).
[…]
