### 2024-02-18 #5884 @ashleychontos
Yes, sorry I was not more clear @MichielsenM.

One of the most important benefits of using docstrings is the ability to automatically generate documentation as an API (see API of [radvel](http://radvel.readthedocs.io/) for an example). Currently it appears as though most (if not all) of the documentation is written out explicitly. I would say this is not of critical importance now but if you end up changing any of the core functionality, you will also have to change this by hand in the documentation. It could save you a lot of time in the long run by generating a code library directly from the docstrings, which will recompile and update any time there is a change to the GitHub repo.

Please let me know if anything I've said is still not clear.
