### 2024-02-22 22cf3f1
docstrings in numpy format
