+++ tests/test_maximum_likelihood_estimator.py
@@ -4,11 +4,11 @@ import unittest
 
 def test_merit_chi2():
     """ test chi2 merit function"""
-    YObs  = np.asarray([1.0, 2.5, 3.7])
-    ObsErr= np.asarray([0.1, 0.05, 0.3])
+    Yobs  = np.asarray([1.0, 2.5, 3.7])
+    obs_err= np.asarray([0.1, 0.05, 0.3])
     YTheo = np.asarray([[1, 3, 4.6], [3, 2.1, 4.9], [1.0, 2.5, 4]])
 
-    result = mle.merit_chi2(YObs, ObsErr, YTheo)
+    result = mle.merit_chi2(Yobs, obs_err, YTheo)
     expected = np.asarray([109, 480, 1])
 
     assert all ([abs((a-b)/a)< 1E-10 for a,b in zip(expected, result)])
@@ -16,11 +16,11 @@ def test_merit_chi2():
 
 def test_merit_mahalanobis():
     """ test MD merit function"""
-    YObs  = np.asarray([1.0, 2.5, 3.7])
-    ObsErr= np.asarray([0.1, 0.05, 0.3])
+    Yobs  = np.asarray([1.0, 2.5, 3.7])
+    obs_err= np.asarray([0.1, 0.05, 0.3])
     YTheo = np.asarray([[1, 3, 4.6], [3, 2.1, 4.9], [1.0, 2.5, 4]])
 
-    result = mle.merit_mahalanobis(YObs, ObsErr, YTheo, generate_output=False)
+    result = mle.merit_mahalanobis(Yobs, obs_err, YTheo, generate_output=False)
     expected = np.asarray([4.96270396, 4.997669, 0.91919192])
 
     assert all ([abs((a-b)/a)< 1E-8 for a,b in zip(expected, result)])        
