+++ foam/pipeline/pipe2_calculate_likelihood.py
@@ -5,11 +5,11 @@ import multiprocessing, os
 from pathlib import Path
 from foam.pipeline.pipeline_config import config
 ###############################################################################
-DataOutDir = Path(f'{os.getcwd()}/meritvalues')
-Path(DataOutDir).mkdir(parents=True, exist_ok=True)
+data_out_dir = Path(f'{os.getcwd()}/meritvalues')
+Path(data_out_dir).mkdir(parents=True, exist_ok=True)
 
-file_Path = Path(f'V_matrix/{config.star}_determinant_conditionNr.tsv')
-if file_Path.is_file(): file_Path.unlink()  #remove file if it exists to avoid duplicate entries on successive runs
+file_path = Path(f'V_matrix/{config.star}_determinant_conditionNr.tsv')
+if file_path.is_file(): file_path.unlink()  #remove file if it exists to avoid duplicate entries on successive runs
 args = []
 observables = []
 for grid in config.grids:
@@ -21,12 +21,12 @@ for grid in config.grids:
                 elif obs == 'f':
                     observed_quantity = 'frequency'
 
-                Theo_path = f'{config.main_directory}/extracted_freqs/surface+{observed_quantity}_{config.star}_{grid}_{method}.hdf'
+                theory_path = f'{config.main_directory}/extracted_freqs/surface+{observed_quantity}_{config.star}_{grid}_{method}.hdf'
                 observables = [obs]
                 if config.observable_additional is not None:
                     observables += config.observable_additional
-                args.append(( Theo_path, observables, merit_function ))
+                args.append(( theory_path, observables, merit_function ))
 
 with multiprocessing.Pool(config.nr_cpu) as p:
-    func = partial(mle.calculate_likelihood, Obs_path=config.observations, star_name=config.star, fixed_params=config.fixed_parameters, grid_parameters=config.grid_parameters)
+    func = partial(mle.calculate_likelihood, obs_path=config.observations, star_name=config.star, fixed_params=config.fixed_parameters, grid_parameters=config.grid_parameters)
     p.starmap(func, args)
