+++ foam/maximum_likelihood_estimator.py
@@ -11,7 +11,7 @@ from foam import build_optimised_pattern as bop
 logger = logging.getLogger('logger.mle_estimator')  # Make a child logger of "logger" made in the top level script
 
 # %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
-def calculate_likelihood(Theo_file, observables=None, merit_function=None, Obs_path=None, star_name=None, fixed_params=None, grid_parameters=None):
+def calculate_likelihood(theory_file, observables=None, merit_function=None, obs_path=None, star_name=None, fixed_params=None, grid_parameters=None):
     """
     Perform a maximum likelihood estimation using the provided type of merit function on the list of observables.
     Writes a data file with the values of the merit function and input parameters of each model.
@@ -19,10 +19,10 @@ def calculate_likelihood(Theo_file, observables=None, merit_function=None, Obs_p
     
     Parameters
     ----------
-    Obs_path: string
+    obs_path: string
         Path to the tsv file with observations, with a column for each observable and each set of errors.
         Column names specify the observable, and "_err" suffix denotes that it's the error.
-    Theo_file: string
+    theory_file: string
         Path to the hdf5 file with the theoretical model input parameters (first set of columns), frequency or period values (last set of columns),
         and possibly extra columns with additional observables (these columns should be in between the input parameters and frequency columns).
     observables: list of strings
@@ -44,44 +44,44 @@ def calculate_likelihood(Theo_file, observables=None, merit_function=None, Obs_p
     elif 'P' or 'dP' in observables:
         observed_quantity = 'period'
     # Read in the observed data and make an array of the observed observables
-    Obs_dFrame = pd.read_table(Obs_path, delim_whitespace=True, header=0, index_col='index')
-    Obs, ObsErr, file_suffix_observables = create_obs_observables_array(Obs_dFrame, observables)
+    obs_dataframe = pd.read_table(obs_path, delim_whitespace=True, header=0, index_col='index')
+    obs, obs_err, file_suffix_observables = create_obs_observables_array(obs_dataframe, observables)
 
     # set the name of the output file
-    _, tail = sf.split_line(Path(Theo_file).stem, star_name)
+    _, tail = sf.split_line(Path(theory_file).stem, star_name)
     filename = f'{star_name}{tail}_{merit_function}_{file_suffix_observables}'
 
     # Theoretical grid data
-    Theo_dFrame = sf.get_subgrid_dataframe(Theo_file,fixed_params)
+    theory_dataframe = sf.get_subgrid_dataframe(theory_file,fixed_params)
 
-    missing_absolute = np.where(Theo_dFrame.columns.to_series().str.contains(f'{observed_quantity}_missing'))[0]    # get the interruptions in the pattern, absolute index in dataframe
-    missing_relative = np.where(Theo_dFrame.filter(like=f'{observed_quantity}').columns.to_series().str.contains(f'{observed_quantity}_missing'))[0]  # get the interruptions in the pattern, index relative within pulsations
-    Theo_dFrame = Theo_dFrame.drop(columns=Theo_dFrame.columns[missing_absolute])   # Remove columns of missing frequencies
+    missing_absolute = np.where(theory_dataframe.columns.to_series().str.contains(f'{observed_quantity}_missing'))[0]    # get the interruptions in the pattern, absolute index in dataframe
[…]
