+++ .github/workflows/deploy-pages.yml
@@ -2,6 +2,7 @@
 name: Deploy Jekyll with GitHub Pages dependencies preinstalled
 
 on:
+  workflow_call:
   push:
     branches:
       - master
