+++ pyproject.toml
@@ -14,9 +14,15 @@ lmfit = "^1.0.3"
 astropy = "^5.1.1"
 hdfdict = "^0.3.1"
 
-[tool.poetry.dev-dependencies]
+[tool.poetry.group.dev.dependencies]
 pytest = "^5.2"
 tomli = "^2.0.1"
+
+[tool.poetry.group.docs]
+optional = true
+
+[tool.poetry.group.docs.dependencies]
+pdoc = "^14.4.0"
 
 [build-system]
 requires = ["poetry-core>=1.0.0"]
