+++ foam/__init__.py
@@ -1,3 +1,13 @@
-__version__ = '1.0.0'
-__all__ = ['additional_constraints','build_optimised_pattern','functions_for_gyre','functions_for_mesa','gmode_rotation_scaling','maximum_likelihood_estimator','model_grid','plot_tools','support_functions']
+__version__ = "1.0.0"
+__all__ = [
+    "additional_constraints",
+    "build_optimised_pattern",
+    "functions_for_gyre",
+    "functions_for_mesa",
+    "gmode_rotation_scaling",
+    "maximum_likelihood_estimator",
+    "model_grid",
+    "plot_tools",
+    "support_functions",
+]
 __docformat__ = "numpy"
