+++ foam/gmode_rotation_scaling.py
@@ -1,15 +1,19 @@
 """Calculate g-mode period spacing patterns in the asymptotic regime using the TAR."""
+
 # Class from T. Van Reeth
-import numpy as np
-import astropy.units as u
 import sys
+
+import astropy.units as u
+import numpy as np
+
+
 ################################################################################
 class Asymptotic(object):
     """
     A python class to calculate g-mode period spacing patterns in the asymptotic regime using the TAR.
     """
 
-    def __init__(self,gyre_dir,kval=0,mval=1,nmin=1,nmax=150):
+    def __init__(self, gyre_dir, kval=0, mval=1, nmin=1, nmax=150):
         """
         Setting up the environment/asymptotic object to calculate g-mode period spacing patterns.
 
@@ -31,13 +35,13 @@ class Asymptotic(object):
 
         self.kval = int(kval)
         self.mval = int(mval)
-        self.nvals = np.arange(nmin,nmax+0.1,1.)
+        self.nvals = np.arange(nmin, nmax + 0.1, 1.0)
 
-        self.lam_fun  = self._retrieve_laplacegrid(gyre_dir)
+        self.lam_fun = self._retrieve_laplacegrid(gyre_dir)
         self.spin, self.lam, self.spinsqlam = self._sample_laplacegrid()
 
-################################################################################
-    def _retrieve_laplacegrid(self,gyre_dir):
+    ################################################################################
+    def _retrieve_laplacegrid(self, gyre_dir):
         """
         Retrieving the function lambda(nu) given in GYRE.
 
@@ -53,28 +57,28 @@ class Asymptotic(object):
             A function to calculate lambda, given spin parameter values as input.
         """
 
-        if(self.kval >= 0):
-            kstr = f'+{self.kval}'
+        if self.kval >= 0:
+            kstr = f"+{self.kval}"
         else:
-            kstr = f'{self.kval}'
-        if(self.mval >= 0):
-            mstr = f'+{self.mval}'
+            kstr = f"{self.kval}"
+        if self.mval >= 0:
+            mstr = f"+{self.mval}"
         else:
-            mstr = f'{self.mval}'
+            mstr = f"{self.mval}"
 
-        infile = f'{gyre_dir}/data/tar/tar_fit.m{mstr}.k{kstr}.h5'
+        infile = f"{gyre_dir}/data/tar/tar_fit.m{mstr}.k{kstr}.h5"
 
-        sys.path.append(gyre_dir+'/src/tar/')
-        import gyre_tar_fit
+        sys.path.append(gyre_dir + "/src/tar/")
         import gyre_cheb_fit
+        import gyre_tar_fit
 
         tf = gyre_tar_fit.TarFit.load(infile)
         lam_fun = np.vectorize(tf.lam)
 
         return lam_fun
 
-################################################################################
-    def _sample_laplacegrid(self,spinmax=1000.,spindensity=1.):
+    ################################################################################
+    def _sample_laplacegrid(self, spinmax=1000.0, spindensity=1.0):
         """
         Sampling the function lambda(nu) that was set up in _retrieve_laplacegrid().
         This subroutine includes a har-coded custom sampling density function for the spin parameter.
@@ -97,43 +101,137 @@ class Asymptotic(object):
             The lambda eigenvalues corresponding to the spin parameter values in the array 'spin'
         spinsqlam: numpy array, dtype=float
             = spin * sqrt(lam)
-            """
+        """
 
-        if((self.kval >= 0) & (self.mval != 0)):
+        if (self.kval >= 0) & (self.mval != 0):
             spinmin = -0.1
             # Relatively "ad hoc" optimal sampling (based on "experience")
[…]
