### 2024-03-24 #5884 @MichielsenM
I've setup the automated API generation and integrated it into the online documentation.
