### 2024-03-25 #5884 @AnkitBarik
@MichielsenM The manuscript as well as the documentation looks much clearer and better than before. Thank you for making these corrections. From my end, I am happy to recommend its publication in JOSS.
