### 2024-03-25 #5884 @AnkitBarik
@MichielsenM @warrickball @ashleychontos Apologies for deleting my previous comment, I quickly realized that I had not actually tested the code functionality yet. I am happy with the revised documentation and manuscript. I am happy to recommend the manuscript for publication. Just one minor comment, the code prints `Calculating liklihoods` , please fix it to 'likelihoods'. :)
