+++ example_setup/KIC7760680_missing_freq/pipeline_missing_freq.py
@@ -15,8 +15,8 @@ if __name__ == "__main__":
         grid_parent_directory=None,  # change to '/YOUR_PATH/MESA_grid' if you need to perform step 0
         grids=["DO"],  # Diffusive Overshooting grid
         rotation_gyre=0.4805,
-        N_periods=32,
-        N_pattern_parts=4,
+        n_periods=32,
+        n_pattern_parts=4,
         pattern_methods=[
             "provided-pulsation",
             "highest-frequency",
