### 2024-04-15 #5884 @warrickball
Hi @ashleychontos, have you had a chance to revisit FOAM following the changes in the review? You've commented on the [manuscript](https://github.com/openjournals/joss-reviews/issues/5884#issuecomment-1830420455) and [documentation](https://github.com/openjournals/joss-reviews/issues/5884#issuecomment-1830501255) (which now includes API docs) but not the code functionality.
