### 2024-04-15 #5884 @ashleychontos
really minor question/point @MichielsenM:

I noticed the documentation states that `FOAM` "requires python 3.9 or newer" under installation but that the project.toml file specifies >=3.9,<3.11. I use Python 3.11.7 -- so I got that error when trying a `pip install .`. Is there a compatibility issue or is this just not well tested? Either way, probably best to make sure it's consistent in both places.
