### 2024-04-16 #5884 @ashleychontos
Some additional notes @MichielsenM:
 - The manuscript looks great! That said, it's still not entirely clear to me who the target audience is for the software package. Upon testing the code functionality, the installation instructions appear to favor developer (i.e. editable) mode. You clearly state the need for an open-source package for modeling gravity modes, which I fully agree with and was surprised to learn that there wasn't one already. Therefore, I would assume that the target audience would likely newer students and other early career researchers with limited experience in both science and software... not developers.
 - Given that you recently moved on from academia, what are the long-term goals for this package? For example, will it still be maintained by you or some collaborator? are you hoping someone else can take over? I didn't see anything about this aspect in the checklist but still seems important for new packages that are published. @warrickball is this relevant in the standard review process? (happy to remove this comment if it isn't)
 - First word under Installation in documentation: please change "Pleas" to "Please" (😆)
