### 2024-04-21 c673e91
updated dependencies, python 3.10-3.12 compatibility
