+++ pyproject.toml
@@ -5,17 +5,18 @@ description = ""
 authors = ["MichielsenM <mathias.michielsen@hotmail.com>"]
 
 [tool.poetry.dependencies]
-python = "^3.9,<3.11"
+python = "^3.9"
 numpy = "^1.21.2"
-h5py = "^3.4.0"
+h5py = "^3.11.0"
 matplotlib = "^3.4.3"
-pandas = {extras = ["hdf5"], version = "^2.0.0"}
+pandas = {extras = ["hdf5"], version = "^2.2.0"}
 lmfit = "^1.0.3"
-astropy = "^5.1.1"
+astropy = "^5.3.3"
 hdfdict = "^0.3.1"
+pyyaml = "^5.1,<5.4"
 
 [tool.poetry.group.dev.dependencies]
-pytest = "^5.2"
+pytest = "^8"
 tomli = "^2.0.1"
 
 [tool.poetry.group.docs]
