+++ pyproject.toml
@@ -13,7 +13,7 @@ pandas = {extras = ["hdf5"], version = "^2.2.0"}
 lmfit = "^1.0.3"
 astropy = "^5.3.3"
 hdfdict = "^0.3.1"
-pyyaml = "^5.1,<5.4"
+pyyaml = "^5.4"
 
 [tool.poetry.group.dev.dependencies]
 pytest = "^8"
