### 2024-04-22 #5884 @MichielsenM
> really minor question/point @MichielsenM:
>
> I noticed the documentation states that `FOAM` "requires python 3.9 or newer" under installation but that the project.toml file specifies >=3.9,<3.11. I use Python 3.11.7 -- so I got that error when trying a `pip install .`. Is there a compatibility issue or is this just not well tested? Either way, probably best to make sure it's consistent in both places.

Hi @ashleychontos
If I remember correctly, I wrote that part when python 3.11 was not released yet, but I somehow got installation problems when specifying >=3.9 without <3.11. I've now updated the dependencies, and verified compatibility (and functionality) for versions 3.9, 3.10, 3.11, and 3.12.
