### 2024-04-22 318c66e
updated installation instructions for non-editable mode
