### 2024-04-22 #5884 @MichielsenM
I've corrected the typo, and updated the installation instructions to give a clearer alternative installation in non-editable mode (using just pip without poetry). This will make  the installation procedure easier for non-developers who don't need an editable installation.
As for the long-term goals of the package, I do plan to provide some maintenance for the package, but I don't plan to actively develop new features by myself. However, I am happy to support people who would contact me wanting to do so.
