### 2024-04-22 d29912e
update installation instructions
