### 2024-05-02 #5884 @warrickball
> * Given that you recently moved on from academia, what are the long-term goals for this package? For example, will it still be maintained by you or some collaborator? are you hoping someone else can take over? I didn't see anything about this aspect in the checklist but still seems important for new packages that are published. @warrickball is this relevant in the standard review process? (happy to remove this comment if it isn't)

@ashleychontos, this is an interesting question. JOSS doesn't require the paper or repo explicitly say anything about how the code will be maintained, nor do we want reviewers to evaluate whether a code looks like it might not be maintained. This isn't at all specific to FOAM. A similar situation arises with e.g. "PhDware": software produced during a PhD that is unmaintained once the student graduates or shortly thereafter.

It looks like @MichielsenM has responded to your other points so let me know if there's anything else you think needs to be done.
