### 2024-05-14 #5884 @ashleychontos
Nothing further @warrickball I have since tested the code installation/functionality/performance and everything checks out. I am happy to recommend the manuscript (and package) for publication in JOSS!
