### 2024-05-15 0858bb7
Some editorial tweaks for JOSS
