+++ paper/paper.md
@@ -3,12 +3,12 @@ title: 'Foam: A Python package for forward asteroseismic modelling of gravity mo
 tags:
   - Python
   - astronomy
-  - stellar astrophysics  
+  - stellar astrophysics
   - asteroseismology
 authors:
   - name: Mathias Michielsen
     orcid: 0000-0001-9097-3655
-    affiliation: 1 
+    affiliation: 1
 
 affiliations:
  - name: Institute of Astronomy, KU Leuven, Celestijnenlaan 200D, B-3001 Leuven, Belgium
@@ -20,7 +20,7 @@ bibliography: paper.bib
 # Summary
 
 Asteroseismology, the study of stellar pulsations, offers insights into the internal structures and evolution of stars. Analysing the variations in a star's brightness allows the determination of fundamental properties such as mass, radius, age, and chemical composition. Asteroseismology heavily relies on computational tools, but a significant number of them are closed-source, thus inaccessible to the broader astronomic community.
-This manuscript presents `Foam`, a python package designed to perform forward asteroseismic modelling of stars exhibiting gravity modes. It automates and streamlines a considerable fraction of the modelling process, comparing grids of theoretical stellar models and their oscillation frequencies to observed frequency sets in stars. 
+This manuscript presents `Foam`, a Python package designed to perform forward asteroseismic modelling of stars exhibiting gravity modes. It automates and streamlines a considerable fraction of the modelling process, comparing grids of theoretical stellar models and their oscillation frequencies to observed frequency sets in stars.
 
 `Foam` offers the flexibility to employ diverse modelling approaches, allowing users to choose different methodologies for matching theoretically predicted oscillations to observations. It provides options to utilise various sets of observables for comparison with their theoretical counterparts, employ different merit functions for assessing goodness of fit, and to incorporate nested subgrids in a statistically rigorous manner. For applications of these methodologies in modelling observed gravity modes, refer to @Michielsen2021 and @Michielsen2023.
 
@@ -40,9 +40,9 @@ Some tools have been developed and made publicly available to model and determin
 
 # Software package overview
 
-`Foam` is designed as a customisable pipeline. It will match theoretical models to observations, computing the goodness of fit of each model based on the selected merit function. Afterwards it will determine the best model alongside the uncertainty region of this solution based on statistical criteria. On the observational side, it will take a list of frequencies as an input, optionally complemented by additional information such as a set of surface properties (effective temperature, surface gravity, luminosity, element surface abundances...). On the theoretical side `Foam` will use a grid of theoretical stellar models, calculated by the user to suit their specific needs. Although the current implementation is made for a grid of stellar equilibrium models computed by `MESA`[@Paxton2011;@Paxton2013;@Paxton2015;@Paxton2018;@Paxton2019;@Jermyn2023], whose pulsation frequencies are computed with `GYRE`[@Townsend2013;@Townsend2018], the majority of the code is not inherently dependent on `MESA`. By making certain adjustments to the modelling pipeline, `Foam` could potentially employ grids generated by different stellar evolution codes. Some suggestions how to approach this are given in the description of [the theoretical model grid](https://michielsenm.github.io/FOAM/Walkthrough) in the online documentation. However, the implementation of such functionality currently remains out of the scope of the project.
[…]
