### 2024-05-15 84f7fba
add two missing DOIs
