+++ paper/paper.bib
@@ -55,7 +55,8 @@ archivePrefix = {arXiv},
 	title = "{Asteroseismology}",
 	booktitle = {Asteroseismology, Astronomy and Astrophysics Library. ISBN 978-1-4020-5178-4. Springer Science+Business Media B.V., 2010, p.},
 	year = "2010",
-	adsurl = {https://ui.adsabs.harvard.edu/\#abs/2010aste.book.....A},
+	doi = {10.1007/978-1-4020-5803-5},
+	adsurl = {https://ui.adsabs.harvard.edu/abs/2010aste.book.....A},
 	adsnote = {Provided by the SAO/NASA Astrophysics Data System},
 	publisher = {Springer, Astronomy and Astrophysics Library}
 }
@@ -375,4 +376,5 @@ archivePrefix = {arXiv},
 	author = {{Claeskens}, G. and {Hjort}, N. L.},
 	title = "{Model Selection and Model Averaging, Cambridge Series in Statistical and Probabilistic Mathematics}",
 	year = 2008,
+	doi = {10.1017/CBO9780511790485}
 }
