### 2024-05-21 #5884 @MichielsenM
Hi @warrickball
I've made a release of version V1.0.0, and archived the release on zenodo https://doi.org/10.5281/zenodo.11237627.
I've also tried to find a DOI for the "Slowly pulsating B stars" reference, but I haven't been able to find one.
Let me know if I need to do anything else.
