### 2024-05-21 #5884 @warrickball
Thanks @MichielsenM! I think that's everything done and I'm about to `recommend-accept`.

> I've also tried to find a DOI for the "Slowly pulsating B stars" reference, but I haven't been able to find one.

Yeah, I also scrounged around to try to find a DOI and didn't manage. It seems A&A hasn't minted them for articles before 2000. Thanks for having another look.
