### 2024-05-25 9f18223
Fix astronomy journal macros
