+++ paper/paper.bib
@@ -19,7 +19,7 @@
 @ARTICLE{Michielsen2023,
        author = {{Michielsen}, M. and {Van Reeth}, T. and {Tkachenko}, A. and {Aerts}, C.},
         title = "{Probing the physics in the core boundary layers of the double-lined B-type binary KIC 4930889 from its gravito-inertial modes}",
-      journal = {\aap},
+      journal = {Astronomy and Astrophysics},
      keywords = {asteroseismology, stars: oscillations, convection, stars: interiors, methods: statistical, techniques: photometric, Astrophysics - Solar and Stellar Astrophysics},
          year = 2023,
         month = nov,
@@ -101,7 +101,7 @@ archivePrefix = {arXiv},
 @ARTICLE{AguirreBorsenKoch2022,
        author = {{Aguirre B{\o}rsen-Koch}, V. and {R{\o}rsted}, J.~L. and {Justesen}, A.~B. and {Stokholm}, A. and {Verma}, K. and {Winther}, M.~L. and {Knudstrup}, E. and {Nielsen}, K.~B. and {Sahlholdt}, C. and {Larsen}, J.~R. and {Cassisi}, S. and {Serenelli}, A.~M. and {Casagrande}, L. and {Christensen-Dalsgaard}, J. and {Davies}, G.~R. and {Ferguson}, J.~W. and {Lund}, M.~N. and {Weiss}, A. and {White}, T.~R.},
         title = "{The BAyesian STellar algorithm (BASTA): a fitting tool for stellar studies, asteroseismology, exoplanets, and Galactic archaeology}",
-      journal = {\mnras},
+      journal = {Monthly Notices of the RAS},
      keywords = {asteroseismology, methods: numerical, methods: statistical, stars: fundamental parameters, Astrophysics - Solar and Stellar Astrophysics, Astrophysics - Earth and Planetary Astrophysics, Astrophysics - Astrophysics of Galaxies},
          year = 2022,
         month = jan,
@@ -119,7 +119,7 @@ archivePrefix = {arXiv},
 @ARTICLE{Rendle2019,
        author = {{Rendle}, Ben M. and {Buldgen}, Ga{\"e}l and {Miglio}, Andrea and {Reese}, Daniel and {Noels}, Arlette and {Davies}, Guy R. and {Campante}, Tiago L. and {Chaplin}, William J. and {Lund}, Mikkel N. and {Kuszlewicz}, James S. and {Scott}, Laura J.~A. and {Scuflaire}, Richard and {Ball}, Warrick H. and {Smetana}, Jiri and {Nsamba}, Benard},
         title = "{AIMS - a new tool for stellar parameter determinations using asteroseismic constraints}",
-      journal = {\mnras},
+      journal = {Monthly Notices of the RAS},
      keywords = {stars: fundamental parameters, stars: oscillations, Astrophysics - Solar and Stellar Astrophysics},
          year = 2019,
         month = mar,
@@ -180,7 +180,7 @@ archivePrefix = {arXiv},
 @ARTICLE{Townsend2013,
 	author = {{Townsend}, R.~H.~D. and {Teitler}, S.~A.},
 	title = "{GYRE: an open-source stellar oscillation code based on a new Magnus Multiple Shooting scheme}",
-	journal = {\mnras},
+	journal = {Monthly Notices of the RAS},
 	keywords = {methods: numerical, stars: evolution, stars: interiors, stars: oscillations, stars: variables: general, Astrophysics - Solar and Stellar Astrophysics},
 	year = 2013,
 	month = Nov,
@@ -197,7 +197,7 @@ archivePrefix = {arXiv},
 @ARTICLE{Townsend2018,
[…]
