+++ paper/paper.bib
@@ -216,7 +216,7 @@ archivePrefix = {arXiv},
 	author = {{Paxton}, Bill and {Bildsten}, Lars and {Dotter}, Aaron and
 	{Herwig}, Falk and {Lesaffre}, Pierre and {Timmes}, Frank},
 	title = "{Modules for Experiments in Stellar Astrophysics (MESA)}",
-	journal = {Astrophysical Journal, Supplement},
+	journal = {The Astrophysical Journal Supplement Series},
 	keywords = {methods: numerical, stars: evolution, stars: general, Astrophysics - Solar and Stellar Astrophysics, Astrophysics - Instrumentation and Methods for Astrophysics},
 	year = 2011,
 	month = jan,
@@ -238,7 +238,7 @@ archivePrefix = {arXiv},
 	{Mankovich}, Christopher and {Montgomery}, M.~H. and {Stello}, Dennis and
 	{Timmes}, F.~X. and {Townsend}, Richard},
 	title = "{Modules for Experiments in Stellar Astrophysics (MESA): Planets, Oscillations, Rotation, and Massive Stars}",
-	journal = {Astrophysical Journals},
+	journal = {The Astrophysical Journal Supplement Series},
 	keywords = {asteroseismology, methods: numerical, planets and satellites: physical evolution, stars: evolution, stars: massive, stars: rotation, Astrophysics - Solar and Stellar Astrophysics, Astrophysics - Instrumentation and Methods for Astrophysics},
 	year = 2013,
 	month = sep,
@@ -260,7 +260,7 @@ archivePrefix = {arXiv},
 	{Dessart}, Luc and {Farmer}, R. and {Hu}, H. and {Langer}, N. and
 	{Townsend}, R.~H.~D. and {Townsley}, Dean M. and {Timmes}, F.~X.},
 	title = "{Modules for Experiments in Stellar Astrophysics (MESA): Binaries, Pulsations, and Explosions}",
-	journal = {Astrophysical Journal, Supplement},
+	journal = {The Astrophysical Journal Supplement Series},
 	keywords = {binaries: general, methods: numerical, nuclear reactions, nucleosynthesis, abundances, shock waves, stars: evolution, stars: oscillations, Astrophysics - Solar and Stellar Astrophysics},
 	year = 2015,
 	month = sep,
@@ -284,7 +284,7 @@ archivePrefix = {arXiv},
 	{Sorokina}, Elena and {Thoul}, Anne and {Townsend}, Richard H.~D. and
 	{Timmes}, F.~X.},
 	title = "{Modules for Experiments in Stellar Astrophysics (MESA): Convective Boundaries, Element Diffusion, and Massive Star Explosions}",
-	journal = {Astrophysical Journals},
+	journal = {The Astrophysical Journal Supplement Series},
 	keywords = {convection, diffusion, hydrodynamics, methods: numerical, stars: evolution, supernovae: general, Astrophysics - Solar and Stellar Astrophysics, Astrophysics - High Energy Astrophysical Phenomena},
 	year = 2018,
 	month = feb,
@@ -308,7 +308,7 @@ archivePrefix = {arXiv},
 	{Townsend}, Richard H.~D. and {Wolf}, William M. and {Zhang}, Michael and
 	{Timmes}, F.~X.},
 	title = "{Modules for Experiments in Stellar Astrophysics (MESA): Pulsating Variable Stars, Rotation, Convective Boundaries, and Energy Conservation}",
-	journal = {Astrophysical Journal, Supplement},
+	journal = {The Astrophysical Journal Supplement Series},
[…]
